requirements.txt
Flask==2.0.3
SQLAlchemy==1.4.31
openai==0.27.0
Werkzeug==2.0.3
python-dotenv
